Java hygiene for shindig
* Fix javadoc @param references to renamed variables.
* Remove useless initializers for some variables and restructure some usages.
* Convert several idioms to use isEmpty().
* Use ternary expressions where they make sense.
* use chars for single-character strings.
* use StringUtils#split or Splitter#on instead of String#split regexes.
* use IOUtils#closeQuietly in finally blocks.
* other stylistic cleanups.

git-svn-id: https://svn.apache.org/repos/asf/shindig/trunk@1190036 13f79535-47bb-0310-9956-ffa450edef68

+++ java/gadgets/src/main/java/org/apache/shindig/gadgets/servlet/ModuleCacheKey.java
@@ -109,7 +109,7 @@ final class ModuleCacheKey implements JobCache.Key {
     }
 
     private void hash(Node node) {
-      hash((short) node.getNodeType());
+      hash(node.getNodeType());
       switch (node.getNodeType()) {
         case Node.ATTRIBUTE_NODE:
         case Node.ELEMENT_NODE:
